frontend: Make a workaround fix for standing table scrolling issue
There's an issue where a standings table sometimes scrolls up
unintentionally on animation.
This happens when a pending submit that is visible in a screen is
moved up to offscreen by getting AC, then the scroll position is moved
to the new location of that team's row.

The root cause is unknown yet, and this commit makes a stopgap fix.